Move Setting related classes into config.flexible.setting package
